Outsource color utils
